tests: Remove unused file
test_updater_ng.py is a little archaic (as it uses the static test
repository content from ye olden days). This commit does not change that
but removes an extra file in client cache dir: it is now quite confusing
as it looks a bit like intermediate root caching but is just an unused
file.

This has the nice side effect that tests now longer need to workaround
this extra file.

Signed-off-by: Jussi Kukkonen <jkukkonen@google.com>